# Simplify trace context benchmark test