Remove unnecessary use of the `http.Request` type and the `oteltest` package (part of #2077).

Sanity check showing that the changes do not change the underlying benchmark.

- with changes:

  ```sh
  $ go test -bench=BenchmarkInject
  goos: linux
  goarch: amd64
  pkg: go.opentelemetry.io/otel/propagation
  cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
  BenchmarkInject/SampledSpanContext-8         	 1139677	      1091 ns/op	     256 B/op	      13 allocs/op
  BenchmarkInject/WithoutSpanContext-8         	53320118	        23.80 ns/op	       0 B/op	       0 allocs/op
  PASS
  ```
- without changes on `main`:

  ```sh
  $ go test -bench=BenchmarkInject
  goos: linux
  goarch: amd64
  pkg: go.opentelemetry.io/otel/propagation
  cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
  BenchmarkInject/SampledSpanContext-8         	  925515	      1117 ns/op	     256 B/op	      13 allocs/op
  BenchmarkInject/WithoutSpanContext-8         	51857061	        23.29 ns/op	       0 B/op	       0 allocs/op
  PASS
  ```